+++ src/Atoms/Media/index.ts
@@ -1,3 +1,4 @@
-import { Media } from './Media';
+import { Media, useMedia } from './Media';
 
+export { useMedia };
 export default Media;