+++ src/Organisms/Development/Development.stories.tsx
@@ -69,9 +69,18 @@ storiesOf('Organisms | Development', module)
   .add('Regression: value is non-zero, but rounded value is 0', () => (
     <Display
       items={[
-        { title: '`value = 0.3`', component: <Development value={0.3} decimals={0} /> },
-        { title: '`value = -0.3`', component: <Development value={-0.3} decimals={0} /> },
-        { title: '`value = -0.3`', component: <Development value={-0.3} maximumDecimals={0} /> },
+        {
+          title: '`value = 0.3` and decimals = 0',
+          component: <Development value={0.3} decimals={0} />,
+        },
+        {
+          title: '`value = -0.3 and decimals = 0`',
+          component: <Development value={-0.3} decimals={0} />,
+        },
+        {
+          title: '`value = -0.3 and maximumDecimals = 0`',
+          component: <Development value={-0.3} maximumDecimals={0} />,
+        },
         {
           title: '`value = 0.3`',
           component: (